Moved ignite class def to module and modified main to take test data
